+++ assets/src/hooks/useBoard/types.ts
@@ -1,0 +1,5 @@
+import { Candidate } from "../../types";
+
+export interface SortedCandidates {
+  [key: string]: Candidate[]
+}
